cleanup pxt.functions.image
